Fix division in `get_average_grade` to ensure correct float result by casting `sum` to `float`.

+++ Student.cpp
@@ -65,7 +65,7 @@ float Student::get_average_grade() const {
             count++;
         }
     }
-    return sum / count;
+    return (float)sum / count;
 }
 
 float Student::get_average_grade_from_math() const {
